Add files via upload
This does not work. Avr-GCC does not recognize a lot of the pseudo-opcodes and thus shouts at you. Further, it does not recognize ',' in many an instruction. The current version has some C in it so as to get rid of the unrecognized .equ directives but other problems still persist